NEI Compatibibility & Creative Tab Roll Back
NEI is looking at the iron array length for items with metadata it
appears, so if you initialize an array that is larger than you use NEI
will show false positives.

Blocks are currently being added to the creative ui during the preinit
event, but tabsEBXL wasn't being set till the init event. I don't know
enough about the difference between client and server side forge
initialization yet to write something that keeps them separate, and just
reverted the code so that I could get it to run for the moment.