NEI Compatibility Fixes / Standardization
In trying to track down why NEI was showing an invalid sapling (subblock
7) in my last patch, the naive fix that I put in allowed newer versions
of NEI to attempt to access a out of bounds texture due. I think I have
finally nailed it down with this patch.

From what I was able to track down some of the blocks, but not all were
only returning the subblock info if it was requested from the EbXL
creative tab, so when NEI or anything else asked for the valid subblocks
it was not returning anything meaning that NEI had to try all possible
subblocks. The patch will definitely need someone else's eyes on it, but
I checked my changes against vanilla blocks and a couple other mods to
see how they did things to make sure I was going along the same line.