BiomeSettting update, and some misc code cleanup.
Biomes settings now accepts 2 or three parameters. If only two
parameters are passed the biome will be enabled by default, how ever if
the third parameters is false then it will be disabled by default.

The march has been disabled by default as it is presently bugged. It
will continue to generate in old worlds where it is enabled in the
config but will no longer be enabled by default.