Log turner skipping fix.
When the log turner is used on quarter logs there was the occasional
chance that it would skip over a rotation when switching between block
id's. This was caused by the the quarter logs' use of multiple block ids
and when the block id's were out of order or wrapped around they would
skip the first rotation for a block id.

This closes #299.