More tree generation updates
More stuff in bring trees up to date with vanilla, and added the
killtree command.

Be careful using the kill tree command in heavily forested areas,
particularly in the twlight forest as it will kill all trees/leaves that
are connected to the tree/leaf that you killed via the  command and this
could cause massive server lag.