Log Turner fix + Tree Gen Update
Fix for the log turner on servers.

Cypress, and Japanese Maple Trees will now refuse to grow if their seeds
result in them intersecting other things such as other trees or terrain.
The Japanese Maple Shrub will not be getting this change as it commonly
intersects the ground in generation.

This collision free change increases world gen just a little, but I
think that I should be able to speed things up a little if I remove the
block tests from the leaf generation. This change should make it to the
Bald Cypress and the Rainbow Eucalyptus later today as well.