New Quarter Logs & Sorted items
I still need to add in the block place event for the quarter logs so
that they place properly. similar to the old logs but with some more
improvements so they don't derp as much.

The config file is also moved out of it's own directory, but I need to
copy the settings over from the old config file if it exists.