Temp redwood/sequois fix
The redwood generator from FN was designed to operate differently than
how our trees generate, so the 2x2 trees were generating in different
locations than the saplings from which they grew.

They will need a rewrite to rewrite that splits the trees of based on
size as in the old code it would randomly select the size of tree to
grow and it would not match the pattern of the saplings planted.