Fix flowers in Creative
getSubBlocks is used by creative to get the list of things to show in
the creative tab. BlockType.values() was pulling the values for all
groups and not just the current group. The "type.metadata() != 0 &
type.metadata() != 5" was a temp patch to remove the root and the shrub
from creative.

It is all fixed so that it only removes them from group 0 and are shown
in all the other groups.