Eelgrass
A prototype of the block system for mc 1.7/v4. This plant base has the
possibility for configurable max height, although at the moment. As the
code stands now plants can not grow similar to how the cactus or reads
do, but they can be placed or generated as such.

At the moment there is not to the plants but that could come later.

Closes #37