Keep old block/item id's when upgrading.
Using forges built in block configurations loader had the problem that
it would mark if a block had been requested and when it was called the
second time to actually load the new configuration it would then detect
that block id as already taken and give the block a new Id. This also
resulted in setting a block id to 0 to not allow any other block to set
it's id to 0 either.

The new design ignores all the special code if the blockID is currently
0 or it is doing an upgrade, and just reads straight from the config
file or uses the current value, allow for things to be set to 0 staying
at 0.