Slab Crafting fix & forge block name prefix change
Forge automatically now appends mod id to blocks in the registry,
meaning that all the old names no longer work. This means that the old
worlds may not be able to be imported into 1.7 as the new block names
may not match the old ones that were class based, but this will be for
the better in the long run.

Also the halfslab recipies were being created to early and created
recipies that could crash the game.