Updated .gitignore
I'm assuming that any file in the source tree that contains .class in the file name isn't something you're going to want, and the Forestry API folder has a lot of junk class files. After this would be merged, this bash command I found on Stack Exchange should remove files in the repo that are marked ignore in .gitignore:
```
git ls-files -i --exclude-from=.gitignore | xargs git rm --cached
```
I've never tried it and don't take my word for it, however.