Prevent ProjectE crash from saplings
ProjectE brute force checks' are making assumptions about metadata and
causing issues with us.

The new sapling block assumes that using the bitbask will be enough and
that no one will be calling with metadata that is completely invalid. If
the block type value is invalid we won't try to convert it to a string.